Scripts: removing all config files with `-R` BUG fix up
* -R is tend to remove not `OK.` config files and
  should not remove them

+++ scripts/v2test.sh
@@ -359,8 +359,12 @@ test_config_file(){
             log_result "$1" "  "
         fi
     elif [[ 1 == $_ext_rm_config_file ]]; then
-        rm $1
-        log_result "$1" "rm"
+        if [[ "$_RES" == "OK." ]]; then
+            log_result "$1" "  "
+        else
+            rm $1
+            log_result "$1" "rm"
+        fi
     else
         log_result "$1"
     fi
